% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNa.r
\name{Cor}
\alias{Cor}
\title{Correlation without NA's'}
\usage{
Cor(x, y = NULL, method = "pearson")
}
\arguments{
\item{x:}{a numeric vector, matrix or data frame.
y: ‘NULL’ (default) or a vector, matrix or data frame with compatible dimensions to ‘x’.  The default is equivalent to ‘y = x’ (but more efficient).}
}
\description{
‘Cor’ computes the correlation of ‘x’ and ‘y’ if these are vectors.  If ‘x’ and ‘y’ are matrices then the correlations between the columns of ‘x’ and the columns of ‘y’ are computed.  Missing values are handled by casewise deletion. If there are no complete cases, that gives an error.
}
