% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryCode-data.r
\docType{data}
\name{countryCode}
\alias{countryCode}
\title{Table with ISO-3 (ISO 3166-1 alpha-3)}
\format{
An object of class data.frame
}
\source{
\href{http://www.trucsweb.com/tutoriels/internet/iso_3166/}{Link}
}
\usage{
countryCode
}
\description{
The names and ISO-3 and ISO-2 codes for countries, based on the UN codes, for use in the construction of datasets.
}
\examples{
data(countryCode)
ISO3 <- 'ZMB'
idx <- countryCode[,2] == ISO3
# what's "ZMB" full name?
print(countryCode[idx,1])
View(countryCode)
}
\keyword{datasets}
