% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.r
\name{countryNameShift}
\alias{countryNameShift}
\title{Change the row names of a matrix/data.frame to ISO-3 country codes}
\usage{
countryNameShift(db)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function changes the row names of a matrix/data.frame to ISO-3 country codes
}
\seealso{
Other iso3: 
\code{\link{iso2name}()},
\code{\link{name2iso}()}
}
\concept{iso3}
