% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{df2mat}
\alias{df2mat}
\title{Convert a dataframe to matrix}
\usage{
df2mat(df)
}
\arguments{
\item{df}{A dataframe including at least one column of numeric values}
}
\value{
A matrix made of the numeric columns from the dataframe. The column names associated will be conserved. The row names will be those of the column titled "ISO3", "ISO2", "country", 'countries', 'pays', 'region', 'regions', 'names', 'name', 'id' if it exists.
}
\description{
Convert a dataframe including numeric values to a matrix; non-numeric columns will be deleted.
}
\examples{
Y <- df2mat(dfPolicyRate)

}
\seealso{
Other convert: 
\code{\link{fac2num}()},
\code{\link{long}()},
\code{\link{mat2df}()}
}
\concept{convert}
