% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.r
\name{incorrect.name2iso}
\alias{incorrect.name2iso}
\title{Indicate which entries are not properly converted by the function 'name2iso'}
\usage{
incorrect.name2iso(db)
}
\description{
Indicate which entries are not properly converted by the function 'name2iso'
}
\seealso{
Other iso3: 
\code{\link{countryNameShift}()},
\code{\link{iso2name}()},
\code{\link{name2iso}()}
}
\concept{iso3}
