% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{long}
\alias{long}
\title{Convert a 'wide' to 'long' format}
\usage{
long(Y)
}
\arguments{
\item{Y}{A panel, 'wide' format matrix or data frame. The columns have to represent different periods of time and the rows different entries (countries, regions, people, companies, etc.).}
}
\value{
The 'long' format data frame, with three columns: 'time', 'id' and 'value'.
}
\description{
Convert A panel, 'wide' format matrix or dataframe to a 'long' format data frame.
}
\examples{
Y <- df2mat(dfPolicyRate)
Yl = long(Y)
head(Yl)

}
\seealso{
Other convert: 
\code{\link{df2mat}()},
\code{\link{fac2num}()},
\code{\link{mat2df}()}
}
\concept{convert}
