% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{mat2df}
\alias{mat2df}
\title{Convert a matrix to data frame}
\usage{
mat2df(Y, ...)
}
\arguments{
\item{Y}{A matrix}

\item{...}{An additional set of matrices and vectors. These matrices have to be of similar dimensions than Y. The vectors have to have a similar amount of values than Y or than the number of rows in Y.}
}
\value{
The data frame made of the matrix or the set of inputs. If multiple inputs are provided, the 'long' format data frame will have with two columns + the number of inputs: 'time' and 'id' follow by 'value', 'values.1', 'values.2', etc.
}
\description{
Convert a matrix (or a set of matrices and vectors) with numeric values to a data frame.
If only Y is provided, convert the matrix Y to a dataframe of similar dimensions.
If additional matrices and/or vectors are provided, the output will be a 'long-reshaped' dataframe.
}
\examples{
Y <- df2mat(dfPolicyRate)
df = mat2df(Y)
head(df)

Y <- df2mat(dfPolicyRate)
df = mat2df(Y, dim(Y)[2]:1, Y, 1:dim(Y)[2])
tail(df)

}
\seealso{
Other convert: 
\code{\link{df2mat}()},
\code{\link{fac2num}()},
\code{\link{long}()}
}
\concept{convert}
