% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.r
\name{name2iso}
\alias{name2iso}
\title{Convert a country name (resp. a vector of country names) to a ISO3 code (resp. a vector of ISO3 codes).}
\usage{
name2iso(name)
}
\arguments{
\item{infile}{A (vector of) character(s) containing country names.}
}
\value{
A (vector of) character(s) containing the ISO3 codes.
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
\seealso{
Other iso3: 
\code{\link{countryNameShift}()},
\code{\link{iso2name}()}
}
\concept{iso3}
