% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipShortcuts.r
\name{pause}
\alias{pause}
\title{Pause until 'Enter' is pressed}
\usage{
pause()
}
\value{
Print "Press <Enter> to continue..." to indicate how to resume the execution.
}
\description{
Pause R script until <Enter> is pressed in R, to make some breaks in an .R file.
}
