% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNa.r
\name{Q}
\alias{Q}
\title{Quantile-alpha of all the values present in its arguments.}
\usage{
Q(x, alpha)
}
\arguments{
\item{x}{A vector containing numeric values}

\item{alpha}{A numeric value in \link{0,1} determining the probability. If alpha=0.5, the function returns the median. If alpha=0.25, the function returns quantile 25\%.}
}
\value{
A numeric scalar, the quantile-alpha of the inputs
}
\description{
‘Q’ returns the quantile-alpha of all the values present in its arguments, omitting NA values.
}
