% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNa.r
\name{Q25}
\alias{Q25}
\title{Quantile-25 of all the values present in its arguments.}
\usage{
Q25(x, alpha)
}
\arguments{
\item{x}{A vector containing numeric values}
}
\value{
A numeric scalar, the quantile-25 of the inputs
}
\description{
‘Q25’ returns the quantile-25 of all the values present in its arguments, omitting NA values.
}
