% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNa.r
\name{Q75}
\alias{Q75}
\title{Quantile-75 of all the values present in its arguments.}
\usage{
Q75(x, alpha)
}
\arguments{
\item{x}{A vector containing numeric values}
}
\value{
A numeric scalar, the quantile-75 of the inputs
}
\description{
‘Q75’ returns the quantile-75 of all the values present in its arguments, omitting NA values.
}
