% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNa.r
\name{Sum}
\alias{Sum}
\title{Sum of all the values present in its arguments.}
\usage{
Sum(x)
}
\arguments{
\item{x}{A vector containing numeric values}
}
\value{
A numeric scalar, the sum of the inputs
}
\description{
‘Sum’ returns the sum of all the values present in its arguments, omitting NA values. Useful to use inside apply functions to avoid errors due to NA's.
}
\examples{
x <- matrix(c((1:11)^2,NA), ncol=4)
apply(x,1,Sum)

}
